Helps if i actually cast to int :)

+++ Python/module2.py
@@ -1,6 +1,6 @@
 length = 20
 width = 10
 
-print("The output of the rectangle is " , (length + width) * 2)
-print("The perimeter of the rectangle is " , length * width)
+print("The output of the rectangle is " , int(length + width) * 2)
+print("The perimeter of the rectangle is " , int(length * width))
